Encryption Enabled + Redis Enabled + Finance Encryption Enabled + Finance Module Query + Extra + User + Guest In V2 Version

+++ controllers/Batch/batchController.js
@@ -1328,17 +1328,23 @@ exports.undo = async (req, res) => {
         student.studentBehaviour = preStudent.studentBehaviour;
         student.department = preStudent.department;
         student.institute = preStudent.institute;
+        if(student?.fee_structure){
         var remain_card = await RemainingList.findOne({
-          fee_structure: `${student?.fee_structure}`,
+          $and: [{ fee_structure: `${student?.fee_structure}`}, { student: student?._id }],
         });
+        if(remain_card?._id){
         student.remainingFeeList.pull(remain_card?._id);
+        }
         if (student?.remainingFeeList_count > 0) {
           student.remainingFeeList_count -= 1;
         }
         if (student?.admissionRemainFeeCount >= remain_card?.remaining_fee) {
           student.admissionRemainFeeCount -= remain_card.remaining_fee;
         }
+        if(remain_card?._id){
         await RemainingList.findByIdAndDelete(remain_card?._id);
+        }
+      }
         student.fee_structure = preStudent?.fee_structure;
         await student.save();
         var clsPrev = await Class.findById(student?.studentClass);
@@ -1364,17 +1370,23 @@ exports.undo = async (req, res) => {
         student.studentBehaviour = preStudent.studentBehaviour;
         student.department = preStudent.department;
         student.institute = preStudent.institute;
+        if(student?.fee_structure){
         var remain_card = await RemainingList.findOne({
-          fee_structure: `${student?.fee_structure}`,
+          $and: [{ fee_structure: `${student?.fee_structure}`}, { student: student?._id }],
         });
+        if(remain_card?._id){
         student.remainingFeeList.pull(remain_card?._id);
+        }
         if (student?.remainingFeeList_count > 0) {
           student.remainingFeeList_count -= 1;
         }
         if (student?.admissionRemainFeeCount >= remain_card?.remaining_fee) {
           student.admissionRemainFeeCount -= remain_card.remaining_fee;
         }
+        if(remain_card?._id){
         await RemainingList.findByIdAndDelete(remain_card?._id);
+        }
+      }
         student.fee_structure = preStudent?.fee_structure;
         await student.save();
         var clsPrev = await Class.findById(student?.studentClass);
@@ -1389,3 +1401,4 @@ exports.undo = async (req, res) => {
     console.log(e);
   }
 };
+
